Move calcPosition() to test
